feat: 分组用户管理

+++ fortuneboot-service/src/main/java/com/fortuneboot/service/fortune/FortuneUserGroupRelationService.java
@@ -1,19 +1,20 @@
 package com.fortuneboot.service.fortune;
 
 import cn.hutool.core.bean.BeanUtil;
+import com.fortuneboot.common.enums.fortune.RoleTypeEnum;
 import com.fortuneboot.common.exception.ApiException;
 import com.fortuneboot.common.exception.error.ErrorCode;
 import com.fortuneboot.domain.command.fortune.FortuneUserGroupRelationAddCommand;
-import com.fortuneboot.domain.command.fortune.FortuneUserGroupRelationModifyCommand;
+import com.fortuneboot.domain.command.fortune.FortuneUserGroupRelationInviteCommand;
 import com.fortuneboot.domain.entity.fortune.FortuneUserGroupRelationEntity;
 import com.fortuneboot.domain.entity.system.SysUserEntity;
 import com.fortuneboot.domain.vo.fortune.FortuneUserGroupRelationVo;
 import com.fortuneboot.factory.fortune.FortuneUserGroupRelationFactory;
 import com.fortuneboot.factory.fortune.model.FortuneUserGroupRelationModel;
+import com.fortuneboot.infrastructure.user.AuthenticationUtils;
+import com.fortuneboot.infrastructure.user.web.SystemLoginUser;
 import com.fortuneboot.repository.fortune.FortuneUserGroupRelationRepository;
 import com.fortuneboot.repository.system.SysUserRepository;
-import jakarta.validation.constraints.NotNull;
-import jakarta.validation.constraints.Positive;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -21,6 +22,7 @@ import org.springframework.stereotype.Service;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
+import java.util.function.Function;
 import java.util.stream.Collectors;
 
 /**
@@ -40,35 +42,49 @@ public class FortuneUserGroupRelationService {
 
     private final SysUserRepository userRepository;
 
-    public void addFortuneUserGroupRelation(FortuneUserGroupRelationAddCommand addCommand) {
+    public void add(FortuneUserGroupRelationAddCommand addCommand) {
         FortuneUserGroupRelationModel relationModel = fortuneUserGroupRelationFactory.create();
         relationModel.loadAddCommand(addCommand);
+        relationModel.checkRepeat(addCommand.getUserId());
         relationModel.insert();
     }
 
-    public void modifyFortuneUserGroupRelation(FortuneUserGroupRelationModifyCommand modifyCommand) {
-        FortuneUserGroupRelationModel relationModel = fortuneUserGroupRelationFactory.loadById(modifyCommand.getUserGroupRelationId());
-        relationModel.loadModifyCommand(modifyCommand);
-        relationModel.updateById();
+    public void inviteUser(FortuneUserGroupRelationInviteCommand inviteCommand) {
+        SysUserEntity user = userRepository.getUserByUserName(inviteCommand.getUsername());
+        if (Objects.isNull(user)) {
+            throw new ApiException(ErrorCode.Business.GROUP_USER_NON_EXIST);
+        }
+        FortuneUserGroupRelationModel relationModel = fortuneUserGroupRelationFactory.create();
+        relationModel.setUserId(user.getUserId());
+        relationModel.loadInviteCommand(inviteCommand);
+        relationModel.checkRepeat(inviteCommand.getUsername());
+        relationModel.insert();
     }
 
-    public void removeFortuneUserGroupRelation(Long userGroupRelationId) {
-        FortuneUserGroupRelationModel relationModel = fortuneUserGroupRelationFactory.loadById(userGroupRelationId);
+    public void removeGroupUser(Long id) {
+        FortuneUserGroupRelationModel relationModel = fortuneUserGroupRelationFactory.loadById(id);
         if (relationModel.getDefaultGroup()) {
             throw new ApiException(ErrorCode.Business.GROUP_CANNOT_DELETE_DEFAULT_GROUP);
+        }
+        SystemLoginUser loginUser = AuthenticationUtils.getSystemLoginUser();
+        if (Objects.equals(loginUser.getUserId(), relationModel.getUserId())){
+            throw new ApiException(ErrorCode.Business.GROUP_CANNOT_DELETE_SELF);
         }
         relationModel.deleteById();
     }
 
-    public List<FortuneUserGroupRelationVo> getUserGroupRelationByGroupId(Long groupId) {
-        List<FortuneUserGroupRelationEntity> userGroupRelationEntityList = fortuneUserGroupRelationRepository.getByGroupId(groupId);
-        List<Long> userIds = userGroupRelationEntityList.stream().map(FortuneUserGroupRelationEntity::getUserId).toList();
+    public List<FortuneUserGroupRelationVo> getGroupUser(Long groupId) {
+        List<FortuneUserGroupRelationEntity> userGroupList = fortuneUserGroupRelationRepository.getByGroupId(groupId);
+        List<Long> userIds = userGroupList.stream().map(FortuneUserGroupRelationEntity::getUserId).toList();
         List<SysUserEntity> userEntityList = userRepository.listByIds(userIds);
-        Map<Long, String> userIdMapName = userEntityList.stream().collect(Collectors.toMap(SysUserEntity::getUserId, SysUserEntity::getUsername));
-        return userGroupRelationEntityList.stream().map(item -> {
-            FortuneUserGroupRelationVo fortuneUserGroupRelationVo = BeanUtil.copyProperties(item, FortuneUserGroupRelationVo.class);
-            fortuneUserGroupRelationVo.setUserName(userIdMapName.get(item.getUserId()));
-            return fortuneUserGroupRelationVo;
+        Map<Long, SysUserEntity> userEntityMap = userEntityList.stream().collect(Collectors.toMap(SysUserEntity::getUserId, Function.identity()));
+        return userGroupList.stream().map(item -> {
+            SysUserEntity user = userEntityMap.get(item.getUserId());
+            FortuneUserGroupRelationVo userGroup = BeanUtil.copyProperties(item, FortuneUserGroupRelationVo.class);
+            userGroup.setUsername(user.getUsername());
+            userGroup.setNickname(user.getNickname());
+            userGroup.setRoleTypeDesc(RoleTypeEnum.getDescByValue(userGroup.getRoleType()));
+            return userGroup;
         }).toList();
     }
